feat(quan): 添加圈子关注功能
- 新增圈子关注、取消关注和检查关注状态的接口和页面逻辑
- 实现用户对圈子的关注和取消关注功能
- 在圈子详情页面显示关注按钮和关注状态
- 优化圈子信息的获取和展示

+++ my-project-backend/src/main/java/com/example/entity/dto/QuanUserBarFollows.java
@@ -9,7 +9,7 @@ import lombok.Data;
 @TableName("quan_user_bar_follows")
 public class QuanUserBarFollows extends BaseModel {
     @TableId(type = IdType.AUTO)
-    private Integer id;
-    private Integer userId;
-    private Integer barId;
+    private Long id;
+    private Long userId;
+    private Long barId;
 }
